:art: 产生了一个公开接口bug

+++ src/traits/PoliciesTraits.hpp
@@ -23,7 +23,7 @@
 
 namespace ui::policies
 {
-enum class Size : uint8_t;
+enum class Size : uint16_t;
 enum class LayoutDirection : uint8_t;
 enum class Alignment : uint8_t;
 enum class Play : uint8_t;
